+++ src/Appwrite/Specification/Format/Swagger2.php
@@ -261,7 +261,12 @@ class Swagger2 extends Format
 
             $bodyRequired = [];
 
-            foreach ($route->getParams() as $name => $param) { // Set params
+            $parameters = \array_merge(
+                $route->getParams(),
+                $route->getLabel('sdk.parameters', []),
+            );
+
+            foreach ($parameters as $name => $param) { // Set params
                 /** @var \Utopia\Validator $validator */
                 $validator = (\is_callable($param['validator'])) ? call_user_func_array($param['validator'], $this->app->getResources($param['injections'])) : $param['validator'];
 
@@ -364,6 +369,11 @@ class Swagger2 extends Format
                         $node['type'] = $validator->getType();
                         $node['format'] = 'password';
                         $node['x-example'] = 'password';
+                        break;
+                    case 'Appwrite\Auth\Validator\Phone':
+                        $node['type'] = $validator->getType();
+                        $node['format'] = 'phone';
+                        $node['x-example'] = '+12065550100';
                         break;
                     case 'Utopia\Validator\Range':
                         /** @var \Utopia\Validator\Range $validator */